+++ crates/filter-list-manager/src/filters/parser/filter_contents_provider/mod.rs
@@ -2,9 +2,11 @@ use crate::io::fetch_by_schemes::fetch_by_scheme;
 use crate::io::url_schemes::UrlSchemes;
 use crate::FilterParserError;
 
+mod check_contents_is_filter_contents;
 pub(crate) mod diff_path_provider;
 pub(super) mod io_provider;
 pub(crate) mod string_provider;
+use check_contents_is_filter_contents::check_contents_is_filter_contents;
 
 /// Provides filters contents.
 /// It can provide filter by `root_filter_url` and resolves its includes
@@ -27,4 +29,9 @@ pub(crate) trait FilterContentsProvider {
 
     /// Sets request timeout if it is not set
     fn set_request_timeout_once(&mut self, request_timeout: i32);
+
+    /// Tries to check the filter data to see if it is a filter or something else, using some heuristics
+    fn pre_check_filter_contents(&self, filter_contents: &str) -> Result<(), FilterParserError> {
+        check_contents_is_filter_contents(filter_contents)
+    }
 }