+++ crates/filter-list-manager/src/io/http/mod.rs
@@ -2,7 +2,7 @@ pub mod error;
 
 use self::error::HttpClientError;
 use reqwest::blocking::{Client as BlockingClient, ClientBuilder as BlockingClientBuilder};
-use reqwest::{Client as AsyncClient, ClientBuilder as AsyncClientBuilder};
+use reqwest::{Client as AsyncClient, ClientBuilder as AsyncClientBuilder, StatusCode};
 use std::time::Duration;
 
 /// Just namespace for http client functions
@@ -26,6 +26,9 @@ impl HttpClient {
             .map_err(HttpClientError::make_body_recovery)
     }
 
+    /// Downloads file from remote server.
+    ///
+    /// Note! This method will fail if status_code != 200
     pub(crate) fn sync_get_file_contents(
         url: &String,
         timeout_ms: i32,
@@ -33,12 +36,21 @@ impl HttpClient {
         let client_builder =
             Self::build_blocking_client(timeout_ms).map_err(HttpClientError::make_network)?;
 
-        client_builder
+        let response = client_builder
             .get(url)
             .send()
-            .map_err(HttpClientError::make_network)?
-            .text()
-            .map_err(HttpClientError::make_body_recovery)
+            .map_err(HttpClientError::make_network)?;
+
+        let status = response.status();
+
+        if status != StatusCode::OK {
+            return Err(HttpClientError::make_only_200_strict(
+                status,
+                url.to_owned(),
+            ));
+        }
+
+        response.text().map_err(HttpClientError::make_body_recovery)
     }
 
     fn build_blocking_client(timeout_ms: i32) -> reqwest::Result<BlockingClient> {